need to fix heredoc leaks

+++ src/tools/cmd_utils.c
@@ -6,13 +6,21 @@
 /*   By: gholloco <gwendal.hollocou@orange.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/27 02:57:53 by gholloco          #+#    #+#             */
-/*   Updated: 2024/11/04 01:58:56 by gholloco         ###   ########.fr       */
+/*   Updated: 2024/11/08 14:48:57 by gholloco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
+#include <stdlib.h>
+#include <unistd.h>
 
-bool	get_doc(t_data *data, int fd, char *word)
+void	unlink_close(char *name, int fd)
+{
+	unlink(name);
+	close(fd);
+}
+
+void	get_doc(t_data *data, int fd, char *word)
 {
 	char	*buf;
 	char	*tmp;
@@ -23,7 +31,6 @@ bool	get_doc(t_data *data, int fd, char *word)
 		buf = readline("> ");
 		if (!buf || !ft_strncmp(word, buf, INT_MAX))
 			break ;
-
 		tmp = expand(data, buf, 1);
 		free(buf);
 		buf = tmp;
@@ -33,31 +40,39 @@ bool	get_doc(t_data *data, int fd, char *word)
 	}
 	free(buf);
 	close(fd);
-	return (true);
 }
 
 void	open_here(t_data *data, t_cmd *cmd, char *word)
 {
 	int	fd;
+	int	pid;
+	int	status;	
 
 	fd = open(".heredoc.tmp", O_CREAT | O_WRONLY | O_TRUNC, 0644);
+	if (cmd->infile != -2)
+		close(cmd->infile);
 	if (fd < 0)
 		cmd->infile = -1;
-	if (!get_doc(data, fd, word))
+	pid = fork();
+	if (pid == -1)
+		return unlink_close(".heredoc.tmp", fd);
+	if (pid == 0)
 	{
-		unlink(".heredoc.tmp");
+		signal(SIGINT, SIG_DFL);
+		get_doc(data, fd, word);
+		exit(0);
+	}
+	close(fd);
+	signal(SIGINT, SIG_IGN);
+	waitpid(pid, &status, 0);
+	if (WIFSIGNALED(status) && WTERMSIG(status) == SIGINT)
+	{
+		printf("\n");
 		cmd->infile = -1;
+		return (data->exit_code = 130, unlink_close(".heredoc.tmp", fd));
 	}
 	fd = open(".heredoc.tmp", O_RDONLY);
 	if (fd > 0)
 		unlink(".heredoc.tmp");
-
-	// a delete un jour
-	// char *buff;
-	// while (read(fd, buff, 10000))
-	// {
-	// 	printf("%s\n", buff);
-	// }
-
 	cmd->infile = fd;
 }
